+++ lib/screens/CheckOutScreen/GroceryCheckOutScreen.dart
@@ -21,6 +21,7 @@ import 'package:qfoods/widgets/ShimmerContainer.dart';
 
 import 'package:http/http.dart' as http;
 import 'package:scroll_to_index/scroll_to_index.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 class GroceryCheckOutScreen extends StatefulWidget {
  final CheckOutModel? checkOut;
@@ -65,6 +66,9 @@ void dispose(){
 }
 
 Future<void> ApplyCouponHandler() async{
+   final SharedPreferences prefs = await SharedPreferences.getInstance();
+    final user_id = await prefs.getInt("qfoods_user_id") ?? null;
+  if(user_id == null) return;
   try{
     context.loaderOverlay.show();
 
@@ -75,7 +79,7 @@ Future<void> ApplyCouponHandler() async{
 
 
     dynamic data = {
-      "user_id": "1",
+      "user_id": user_id?.toString(),
       "latitude": widget.checkOut?.latitude?.toString(),
       "longitude": widget.checkOut?.longitude?.toString(),
       "code": couponController.value.text
@@ -108,6 +112,9 @@ Future<void> ApplyCouponHandler() async{
 
 
 Future<void> CreateOrderHandler() async{
+   final SharedPreferences prefs = await SharedPreferences.getInstance();
+    final user_id = await prefs.getInt("qfoods_user_id") ?? null;
+  if(user_id == null) return;
   if(payment_id == 0){
     await controller.scrollToIndex(2,
         preferPosition: AutoScrollPosition.begin);
@@ -127,7 +134,8 @@ payment_method_error = false;
   });
 try{
   final data = {
-"user_id": "1", "latitude": widget?.checkOut?.latitude?.toString() ?? "0",
+"user_id": user_id?.toString(),
+ "latitude": widget?.checkOut?.latitude?.toString() ?? "0",
  "longitude": widget?.checkOut?.longitude?.toString() ?? "", "address1": widget?.checkOut?.address1?.toString() ?? "",
   "address2": widget?.checkOut?.address2?.toString() ?? "","landmark": widget?.checkOut?.landMark?.toString() ?? "",
    "alternate_phone_number": widget?.checkOut?.alternate_phone_number?.toString() ?? "", 
@@ -163,13 +171,16 @@ catch(e){
 
 Future<void> CheckOutHandler() async{
   print("object");
+   final SharedPreferences prefs = await SharedPreferences.getInstance();
+    final user_id = await prefs.getInt("qfoods_user_id") ?? null;
+  if(user_id == null) return;
   setState(() {
     loading = true;
   });
 try{
   
 dynamic jsonMap = {
-      "user_id": "1",
+      "user_id": user_id?.toString(),
       "latitude": widget.checkOut?.latitude?.toString(),
       "longitude": widget.checkOut?.longitude?.toString()
     };
@@ -516,9 +527,12 @@ Widget PlaceOrder(){
                                         children:  [
                                               // A SlidableAction can have an icon and/or a label.
                                               SlidableAction(
-                                                onPressed: ((context) {
+                                                onPressed: ((context) async{
+                                                   final SharedPreferences prefs = await SharedPreferences.getInstance();
+    final user_id = await prefs.getInt("qfoods_user_id") ?? null;
+  if(user_id == null) return;
                              DeleteProductHandler({
-                              "user_id": "1",
+                              "user_id": user_id?.toString(),
                               "cart_id": [cartProvider?.CartData?.items?[index]?.cartId]
                             });
                                                 }),
@@ -535,9 +549,12 @@ Widget PlaceOrder(){
                                         children:  [
                                               // A SlidableAction can have an icon and/or a label.
                                               SlidableAction(
-                                                 onPressed: ((context) {
+                                                 onPressed: ((context) async{
+                                                   final SharedPreferences prefs = await SharedPreferences.getInstance();
+    final user_id = await prefs.getInt("qfoods_user_id") ?? null;
+  if(user_id == null) return;
                              DeleteProductHandler({
-                              "user_id": "1",
+                              "user_id": user_id?.toString(),
                               "cart_id": [cartProvider?.CartData?.items?[index]?.cartId]
                             });
                                                 }),
@@ -612,23 +629,28 @@ Widget PlaceOrder(){
                                                mainAxisAlignment: MainAxisAlignment.center,
                                                             children: [
                                                               InkWell(
-                                                                onTap: (){
-                                                                 
+                                                                onTap: () async{
+                                                                  final SharedPreferences prefs = await SharedPreferences.getInstance();
+    final user_id = await prefs.getInt("qfoods_user_id") ?? null;
+  if(user_id == null) return;
                                                                   if(cartProvider?.CartData?.items?[index]?.variantid != null){
                                                         UpdateQuantityHandler({
                                                           "cart_id": cartProvider?.CartData?.items?[index]?.cartId?.toString() ?? "",
                                                           "type": "minus",
-                                                          "user_id": "1",
+                                                          "user_id":user_id?.toString(),
                                                           "product_type": "variant",
                                                         "latitude": widget.checkOut?.latitude?.toString(),
                                                                     "longitude": widget.checkOut?.longitude?.toString()
                                                                    });
                                                       
                                                        }else{
+                                                         final SharedPreferences prefs = await SharedPreferences.getInstance();
+    final user_id = await prefs.getInt("qfoods_user_id") ?? null;
+  if(user_id == null) return;
                                               UpdateQuantityHandler({
                                                           "cart_id": cartProvider?.CartData?.items?[index]?.cartId?.toString() ?? "",
                                                           "type": "minus",
-                                                          "user_id": "1",
+                                                          "user_id": user_id?.toString(),
                                                           "product_type": "",
                                                         "latitude": widget.checkOut?.latitude?.toString(),
                                                                     "longitude": widget.checkOut?.longitude?.toString()
@@ -657,23 +679,28 @@ Widget PlaceOrder(){
                                                                 child: Text("${cartProvider?.CartData?.items?[index]?.quantity ?? ""}", style: TextStyle(color: AppColors.primaryColor, fontFamily: FONT_FAMILY, fontSize: ScreenUtil().setSp(14.0), fontWeight: FontWeight.bold),),
                                                               ),
                                                                 InkWell(
-                                                                    onTap: (){
-                                                                  
+                                                                    onTap: () async{
+                                                                   final SharedPreferences prefs = await SharedPreferences.getInstance();
+    final user_id = await prefs.getInt("qfoods_user_id") ?? null;
+  if(user_id == null) return;
                                                         if(cartProvider?.CartData?.items?[index]?.variantid != null){
                                                         UpdateQuantityHandler({
                                                           "cart_id": cartProvider?.CartData?.items?[index]?.cartId?.toString() ?? "",
                                                           "type": "plus",
-                                                          "user_id": "1",
+                                                          "user_id": user_id?.toString(),
                                                           "product_type": "variant",
                                                         "latitude": widget.checkOut?.latitude?.toString(),
                                                                     "longitude": widget.checkOut?.longitude?.toString()
                                                         });
                                                       
                                                        }else{
+                                                         final SharedPreferences prefs = await SharedPreferences.getInstance();
+    final user_id = await prefs.getInt("qfoods_user_id") ?? null;
+  if(user_id == null) return;
                                               UpdateQuantityHandler({
                                                           "cart_id": cartProvider?.CartData?.items?[index]?.cartId?.toString() ?? "",
                                                           "type": "plus",
-                                                          "user_id": "1",
+                                                          "user_id":user_id?.toString(),
                                                           "product_type": "",
                                                         "latitude": widget.checkOut?.latitude?.toString(),
                                                                     "longitude": widget.checkOut?.longitude?.toString()
@@ -702,9 +729,12 @@ Widget PlaceOrder(){
                                                     ),
                               
                                                                     cartProvider?.CartData?.items?[index]?.status == 0
-                                             ?  IconButton(onPressed: (){
+                                             ?  IconButton(onPressed: () async{
+                                               final SharedPreferences prefs = await SharedPreferences.getInstance();
+    final user_id = await prefs.getInt("qfoods_user_id") ?? null;
+  if(user_id == null) return;
                                                DeleteProductHandler({
-                                                "user_id": "1",
+                                                "user_id": user_id?.toString(),
                                                 "cart_id": [cartProvider?.CartData?.items?[index]?.cartId]
                                               });  
                                              }, icon: Icon(Icons.delete, color: AppColors.greycolor,size: ScreenUtil().setSp(20),))
